+++ Source/Core/Utilities/NovusVariants.pas
@@ -41,13 +41,8 @@ begin
            Result.VUnicodeString := Pointer(U);
          end;
 
-       varOleStr:
-         begin
-            New(Result.VString);
-            Result.VType := vtString;
-            Result.VString^ := VarToStr(aValue);
-         end;
-
+       varOleStr,
+       256,
        vtString:
           begin
             New(Result.VString);
@@ -68,7 +63,7 @@ begin
        vtInt64:         ;
        vtUnicodeString: ;
   else
-      Raise Exception.Create('TNovusVariants.VarToVarRec: Unrecognized variant type:'+ IntToStr(aValue));
+      Raise Exception.Create('TNovusVariants.VarToVarRec: Unrecognized variant type: '+ VarTypeAsText(VarType(aValue)) + ':' + IntToStr(VarType(aValue)));
   end;
 end;